refactor: _를 사용하지 않는 이름으로 변경
