+++ data/src/main/kotlin/com/example/data/repository/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package com.example.data.repository
 
-import com.example.data.data_source.DataSourceModule
+import com.example.data.datasource.DataSourceModule
 import com.example.domain.repository.GitHubRepository
 
 internal object RepositoryModule {
