+++ data/src/main/kotlin/com/example/data/retrofit/RetrofitNetwork.kt
@@ -1,6 +1,6 @@
 package com.example.data.retrofit
 
-import com.example.data.data_source.NetworkDataSource
+import com.example.data.datasource.NetworkDataSource
 import com.example.data.model.NetworkRepositoryResource
 
 internal class RetrofitNetwork(private val retrofitNetworkApi: RetrofitNetworkApi) : NetworkDataSource {
